move rknn code to other module
